review fixes for PR #27 — TOU index toInt + Pro clamps in compare path
Two Codex review comments on commit 0bc9b34:

- tariff.dart `importPriceAtHour` / `exportPriceAtHour` now explicitly
  `.toInt()` the clamped hour index. `int.clamp(int, int)` returns
  `num` statically; strict analyzers reject the resulting `List[num]`
  expression. Matches the existing `SummaryAggregator.bankDaily`
  convention. (P1)

- Extract the run-time Pro-gating logic from `ConfigDraft.buildForRun`
  into a new top-level `applyProGates(SimulationConfig)` helper so
  `scenario_comparison_controller.dart` can apply the same clamps to
  configs it loads from the DB. Without this, a free build comparing
  Pro-authored saved scenarios still executed `simulationYears > 1`
  and the TOU tariff arrays — bypassing the gate that the Auswertung-
  tab Run button enforces. (P2)

+++ pv_calculator_dexus_overlay/packages/pv_engine/lib/src/tariff.dart
@@ -31,7 +31,10 @@ class TariffConfig {
   double importPriceAtHour(double hourOfDay) {
     final hours = hourlyImportPrices;
     if (hours == null) return importPricePerKwh;
-    return hours[hourOfDay.floor().clamp(0, 23)];
+    // `int.clamp(int, int)` returns `num` statically; the explicit
+    // toInt() matches the existing engine convention (see
+    // `SummaryAggregator.bankDaily`) and keeps strict analyzers happy.
+    return hours[hourOfDay.floor().clamp(0, 23).toInt()];
   }
 
   /// Price earned for one kWh exported during a step whose midpoint is
@@ -39,7 +42,7 @@ class TariffConfig {
   double exportPriceAtHour(double hourOfDay) {
     final hours = hourlyExportPrices;
     if (hours == null) return exportPricePerKwh;
-    return hours[hourOfDay.floor().clamp(0, 23)];
+    return hours[hourOfDay.floor().clamp(0, 23).toInt()];
   }
 
   void validate() {
